+++ CHANGELOG.md
@@ -4,7 +4,10 @@ CHANGELOG
 0.4.0
 -----
 
+* dropped suppport for PHP < 5.6 and HHVM
+
 * made the package compatible with `3.x` releases of `ramsey/uuid`
+
 * allow `2.x` and `3.x` releases of the `php-xapi/model` package too
 
 0.3.0